+++ src/app/App.test.tsx
@@ -1,7 +1,8 @@
 import { describe, expect, it } from "vitest";
-import { render, screen } from "@testing-library/react";
 import { MemoryRouter } from "react-router-dom";
-import App from "@/App";
+
+import { render, screen } from "@testing-library/react";
+import App from "./App";
 
 describe("Renders main page correctly", async () => {
   it("Should render the page correctly", async () => {
