+++ src/app/App.tsx
@@ -1,4 +1,5 @@
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
+
 import routerConfig from "@/routes";
 
 const App: React.FC = () => {
