+++ src/app/main.tsx
@@ -1,6 +1,7 @@
 import { createRoot } from "react-dom/client";
-import App from "@/App";
+
 import "@/index.css";
+import App from "./App";
 
 createRoot(document.getElementById("root")!).render(
   // <StrictMode>
