+++ src/components/Header/index.tsx
@@ -1,5 +1,6 @@
-import { COGNITO_LOGIN_URL } from "@/constants/apiConfig";
 import { Link } from "react-router-dom";
+
+import { COGNITO_LOGIN_URL } from "@/constants/apiConfig";
 
 const Header = () => {
   return (
