+++ src/services/apiServices/index.ts
@@ -1,1 +1,2 @@
 export * from "./tourApis";
+export * from "./authApis";
